feat(controller): add support for authenticated external registry.
This patch add ability to use an private image (stored on private docker hub
or aws ECR or quay) as a base for deployment with deis pull command.
For now we can only setup one external registry at a time, the plan in the future
is to support multiple private registry at the same time.

Those etcd keys were added:
  * /deis/controller/private_ext_registry/flavor:
    contains the type of the authentication, two value are
    accepted 'ecr' and 'default'.
      - when ecr is used, we assume we are
        running on aws ec2 and this the instance we are working on has an attahced
        role to issue aws ecr get-login,
        also in this case 'data' would be a json dict containing the region
        where the regitry is configured.
      - when 'default' is used, the data would contain a json dict
        with username, passowrd, registry, and email to use for docker login.

  * /deis/controller/private_ext_registry/data