[CONTROLLER] Remove no sync from controller bin/boot to ensure that the etcd value would be sent after member synchronization
Because it can happen that under load of confd the refresh of the
controller IP in etcd takes more time then the ttl.
Leading to a situation where the controller is not accessible
from the router (router returning 503).

By removing the --no-sync value, we ensure that the etcd value
would be sent after member synchronisation, which means we would
ask for memeber IPs and route the request directly to etcd main
member, to garantee the value is written correctly.