Improve the availability zone distribution logic. We don't rely on a static list of "available" availability zone name. The list is dynamycly created from fleet metadata. The tags used is always named "az" (might be improved by making the tag name a parameter).
So if the machines are tagged with az=X the X would be collected,
and the unit will be distributed across those Xs. If that list
change because of autoscaling event, the list will always be regenerated
before scheduling unit.

If there is no az tag, nothing happen the logic stays the same as before.